Restructure project to MVC
