+++ src/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Product } from './interfaces/product.interface';
+import { Product } from '../type/product.interface';
 import productsData from '../data/products.json';
 export class ProductsService {
   private products: Product[] = productsData;
