+++ src/type/product.interface.ts
